Move Wheel tests
